+++ app/controllers/better_mailer_previews/application_controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module BetterMailerPreviews
   class ApplicationController < ActionController::Base
   end