+++ config/initializers/assets.rb
@@ -1,1 +1,3 @@
-Rails.application.config.assets.precompile += %w(better_mailer_previews/application.css better_mailer_previews/application.js)
+# frozen_string_literal: true
+
+Rails.application.config.assets.precompile += %w[better_mailer_previews/application.css better_mailer_previews/application.js]