+++ lib/better_mailer_previews.rb
@@ -1,5 +1,7 @@
-require "better_mailer_previews/version"
-require "better_mailer_previews/engine"
+# frozen_string_literal: true
+
+require 'better_mailer_previews/version'
+require 'better_mailer_previews/engine'
 
 module BetterMailerPreviews
   # Your code goes here...