+++ spec/dummy/app/mailers/invoice_mailer.rb
@@ -1,0 +1,9 @@
+class InvoiceMailer < ApplicationMailer
+  def saas
+    mail(to: "customer@example.com", subject: "Your SaaS invoice")
+  end
+
+  def basic
+    mail(to: "customer@example.com", subject: "Your invoice")
+  end
+end