+++ spec/dummy/app/mailers/test/test_mailer.rb
@@ -1,0 +1,7 @@
+module Test
+  class TestMailer < ApplicationMailer
+    def github_test
+      mail(to: "customer@example.com", subject: "Namespaced mailer")
+    end
+  end
+end