+++ spec/dummy/spec/mailers/previews/invoice_mailer_preview.rb
@@ -1,0 +1,9 @@
+class InvoiceMailerPreview < ActionMailer::Preview
+  def saas
+    InvoiceMailer.saas
+  end
+
+  def basic
+    InvoiceMailer.basic
+  end
+end