+++ spec/dummy/spec/mailers/previews/test/test_mailer_preview.rb
@@ -1,0 +1,7 @@
+module Test
+  class TestMailerPreview < ActionMailer::Preview
+    def github_test
+      Test::TestMailer.github_test
+    end
+  end
+end